Refactor package structure
This commit refactors the package structure. In particular, the server
package is out-sourced into its own repository, and a proper entry point
is created.
