+++ package.json
@@ -1,17 +1,8 @@
 {
   "dependencies": {
-    "@types/deep-equal": "^1.0.1",
-    "@types/express": "^4.16.0",
-    "@types/morgan": "^1.7.35",
     "@types/node": "^10.5.7",
     "@types/xmldom": "^0.1.29",
-    "body-parser": "^1.18.3",
-    "deep-equal": "^1.0.1",
-    "express": "^4.16.3",
     "jsonschema": "^1.2.4",
-    "morgan": "^1.9.0",
-    "ts-json-schema-generator": "^0.32.0",
-    "ts-node": "^7.0.0",
     "xmldom": "^0.1.27"
   },
   "scripts": {
@@ -31,6 +22,8 @@
   "license": "CC-BY-3.0",
   "devDependencies": {
     "@types/mocha": "^5.2.5",
-    "mocha": "^5.2.0"
+    "mocha": "^5.2.0",
+    "ts-node": "^7.0.0",
+    "ts-json-schema-generator": "^0.32.0"
   }
 }
