+++ server/package.json
@@ -1,0 +1,26 @@
+{
+    "dependencies": {
+      "@types/mocha": "^5.2.5",
+      "@types/xmldom": "^0.1.29",
+      "body-parser": "^1.18.3",
+      "express": "^4.16.3",
+      "jsonschema": "^1.2.4",
+      "mocha": "^5.2.0",
+      "morgan": "^1.9.0",
+      "ts-node": "^7.0.0",
+      "xmldom": "^0.1.27"    
+    },
+    "scripts": {
+      "serve": "ts-node src/index.ts",
+      "serveprod": "ts-node -T src/index.ts"
+    },
+    "private": true,
+    "name": "openmath-json-schema-server",
+    "version": "0.0.1",
+    "description": "An OpenMath JSON Schema server",
+    "main": "src/index.ts",
+    "repository": "git@github.com:tkw1536/OpenMath-JSON.git",
+    "author": "Tom Wiesing <tkw01536@gmail.com>",
+    "license": "CC-BY-3.0"
+}
+
