+++ server/src/api.d.ts
@@ -1,5 +1,5 @@
 import { Validator, ValidatorResult } from "jsonschema";
-import { omany } from "../schema/openmath";
+import { omany } from ",./../schema/openmath";
 
 type IApiResponse<T extends {}> = IApiFailure | IApiSuccess<T>;
