+++ server/src/static/index.html
